# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Palindot API",
        "description": "Api created for a URL Shortener that generates palindromic values.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Palindot API"
        }
    ],
    "paths": {
...
        "/register": {
            "post": {
                "summary": "Registers a new user.",
                "description": "New user registration",
                "operationId": "0903f2e0df1a8857bee538fc055417f0",
                "requestBody": {
                    "description": "Registration parameters",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "summary": "Logs in a user.",
                "description": "Login user",
                "operationId": "2dd578ff7aba2721293dac66833b27d1",
                "requestBody": {
                    "description": "Login parameters",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User logged in successfully"
                    }
                }
            }
        },
...
        "/subsets": {
            "get": {
                "summary": "Gets the list of subsets of the user.",
                "description": "Display a listing of the resource.",
                "operationId": "3fe2eccdbc50d273a44090d1e756f7da",
                "responses": {
                    "200": {
                        "description": "List of subsets"
                    }
                }
            },
            "post": {
                "summary": "Creates a new subset.",
                "description": "Store a newly created resource in storage.",
                "operationId": "fe0673bf6b64610e72c21a5f2aef06b8",
                "requestBody": {
                    "description": "Subset parameters",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SubsetRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Subset created successfully"
                    }
                }
            }
        },
        "/subsets/{id}": {
            "get": {
                "summary": "Gets a specific subset.",
                "description": "Display the specified resource.",
                "operationId": "a99dd0ec45a0f3b9521bc14165d93e35",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the subset",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Specific subset"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                }
            },
            "put": {
                "summary": "Updates a specific subset.",
                "description": "Update the specified resource in storage.",
                "operationId": "28353003e123ba04536c41e361aea92a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the subset",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Subset parameters",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SubsetRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Subset updated successfully"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                }
            },
            "delete": {
                "summary": "Deletes a specific subset.",
                "description": "Remove the specified resource from storage.",
                "operationId": "fbc8b190526c83ca8e70857d3365e8a5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the subset",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subset deleted successfully"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "LoginRequest": {
                "description": "Requires the data necessary to log in",
                "properties": {
                    "email": {
                        "description": "The email of the user",
                        "type": "string"
                    },
                    "password": {
                        "description": "The password of the user",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "RegisterRequest": {
                "description": "Requires the data necessary to register a user",
                "properties": {
                    "name": {
                        "description": "The name of the user",
                        "type": "string"
                    },
                    "email": {
                        "description": "The email of the new user",
                        "type": "string"
                    },
                    "password": {
                        "description": "The password of the new user. It must be confirmed",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "SubsetRequest": {
                "description": "Requires the data necessary to store and update a subset",
                "properties": {
                    "subset_name": {
                        "description": "The name of the subset",
                        "type": "string"
                    },
                    "subset_descr": {
                        "description": "The description of the subset. It can be null and it's not required. Must contain 20 characters at least",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
...
    }
}